+++ R/latex.R
@@ -127,6 +127,11 @@ build_latex_caption <- function(ht, longtable = FALSE) {
   lab <- caption_data$label
   cap <- caption_data$text
 
+  if (is.na(cap) && is.na(lab) &&
+    (caption_data$quarto_caption || caption_data$quarto_label)) {
+    return("")
+  }
+
   if (is.na(cap)) {
     cap <- ""
   } else {
@@ -160,7 +165,8 @@ build_latex_caption <- function(ht, longtable = FALSE) {
 
   lab <- if (is.na(lab) || caption_data$label_in_caption) "" else sprintf("\\label{%s}\n", lab)
   cap <- paste(cap, lab)
-  if (using_quarto(min_version = "1.4") &&
+  if (nzchar(trimws(cap)) &&
+    using_quarto(min_version = "1.4") &&
     getOption(
       "huxtable.knitr_output_format",
       guess_knitr_output_format()